use SemVersion for Package version
