+++ crates/fluvio-controlplane-metadata/Cargo.toml
@@ -24,6 +24,7 @@ thiserror = "1.0.20"
 base64 = "0.13.0"
 bytes = { version = "1" }
 flate2 = { version = "1.0.22", optional = true }
+semver = "1.0"
 
 # External Fluvio dependencies
 fluvio-future = { version = "0.4.0" }
